+++ CHANGELOG.md
@@ -12,10 +12,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - App: Add `tabs` page type.
+- Block: Add support for client-side validation in Form block by setting `required: true`.
 - Server: `$created` and `$updated` are exposed in the API.
 
 ### Changed
 
+- Block: Drop support for `bool` in Form block. Use `boolean` instead.
 - Server: Remove `/_/` routing. Apps are now served from `@organization/appname` and all `/_/`
   routes serve from `/`. This allows for app names no longer being unique for the entirety of
   Appsemble but instead within the organization.